requirements.txt
click==8.1.3
colorama==0.4.4
Flask==2.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
pydantic==1.9.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
typing_extensions==4.2.0
Werkzeug==2.1.2
zipp==3.8.0
python-dotenv==0.20.0